+++ ydb/tools/mnc/cli/ut/test_disks.py
@@ -4,91 +4,8 @@ import types
 import unittest
 from unittest import mock
 
-from ydb.tools.mnc.cli import main
 from ydb.tools.mnc.cli.commands import disks
-from ydb.tools.mnc.lib import agent_client
-from ydb.tools.mnc.lib.exceptions import CliError
-
-
-class Response:
-    def __init__(self, status=200, payload=None, text="error", json_exception=None):
-        self.status = status
-        self.payload = payload or {}
-        self.text_value = text
-        self.json_exception = json_exception
-
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, exc_type, exc, tb):
-        pass
-
-    async def text(self):
-        return self.text_value
-
-    async def json(self):
-        if self.json_exception is not None:
-            raise self.json_exception
-        return self.payload
-
-
-class ClientSession:
-    def __init__(self, response):
-        self.response = response
-        self.post_calls = []
-        self.get_calls = []
-
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, exc_type, exc, tb):
-        pass
-
-    def post(self, url, json):
-        self.post_calls.append((url, json))
-        return self.response
-
-    def get(self, url):
-        self.get_calls.append(url)
-        return self.response
-
-
-class Task:
-    def __init__(self):
-        self.updates = []
-
-    async def update(self, **kwargs):
-        self.updates.append(kwargs)
-
-
-class ParentTask(Task):
-    def __init__(self):
-        super().__init__()
-        self.subtasks = []
-
-    async def add_subtask(self, *args, **kwargs):
-        task = Task()
-        self.subtasks.append((args, kwargs, task))
-        return task
-
-
-class CliMainTest(unittest.IsolatedAsyncioTestCase):
-    async def test_async_main_raises_cli_error_when_command_returns_false(self):
-        async def do(args):
-            return False
-
-        module = types.SimpleNamespace(
-            __name__="ydb.tools.mnc.cli.commands.fail",
-            expected_config=None,
-            add_arguments=lambda parser: None,
-            do=do,
-        )
-
-        with mock.patch.object(main, "modules", [module]), mock.patch("sys.argv", ["mnc", "fail"]):
-            with self.assertRaises(CliError) as error:
-                await main.async_main()
-
-        self.assertEqual(str(error.exception), "Command 'fail' failed")
+from ydb.tools.mnc.cli.ut.helpers import ParentTask
 
 
 class DisksCommandTest(unittest.IsolatedAsyncioTestCase):
@@ -436,64 +353,3 @@ class DisksCommandTest(unittest.IsolatedAsyncioTestCase):
                 [disks.common.Device("label1", "/dev/sdb")],
                 parent_task=ParentTask(),
             ))
-
-
-class AgentClientTest(unittest.IsolatedAsyncioTestCase):
-    async def test_post_json_with_parent_task_returns_response_from_step(self):
-        async def post_json(host, path, payload, port):
-            self.assertEqual((host, path, payload, port), ("host1", "/path", {"key": "value"}, 8999))
-            return {"success": True}
-
-        with mock.patch.object(agent_client, "_post_json", post_json):
-            response = await agent_client.post_json("host1", "/path", {"key": "value"}, parent_task=ParentTask())
-
-        self.assertEqual(response, {"success": True})
-
-    async def test_post_json_with_parent_task_returns_none_on_step_failure(self):
-        parent_task = ParentTask()
-
-        async def post_json(host, path, payload, port):
-            return None
-
-        with mock.patch.object(agent_client, "_post_json", post_json):
-            response = await agent_client.post_json("host1", "/path", {}, parent_task=parent_task)
-
-        self.assertIsNone(response)
-        self.assertEqual(parent_task.subtasks[0][2].updates, [])
-
-    async def test_post_json_returns_none_on_http_error(self):
-        session = ClientSession(Response(status=500, text="boom"))
-
-        with mock.patch.object(agent_client.aiohttp, "ClientSession", lambda: session):
-            self.assertIsNone(await agent_client._post_json("host1", "/path", {"key": "value"}))
-
-        self.assertEqual(session.post_calls, [("http://host1:8999/path", {"key": "value"})])
-
-    async def test_post_json_returns_none_on_invalid_json(self):
-        session = ClientSession(Response(status=200, json_exception=ValueError("bad json")))
-
-        with mock.patch.object(agent_client.aiohttp, "ClientSession", lambda: session):
-            self.assertIsNone(await agent_client._post_json("host1", "/path", {}))
-
-    async def test_health_check_succeeds_when_required_features_enabled(self):
-        session = ClientSession(Response(payload={"enabled_features": ["nodes", "disks"]}))
-
-        with mock.patch.object(agent_client.aiohttp, "ClientSession", lambda: session):
-            self.assertTrue(await agent_client.CheckAgentHealthOnHost("host1").action())
-
-        self.assertEqual(session.get_calls, ["http://host1:8999/health"])
-
-    async def test_health_check_fails_on_missing_features(self):
-        session = ClientSession(Response(payload={"enabled_features": ["nodes"]}))
-
-        with mock.patch.object(agent_client.aiohttp, "ClientSession", lambda: session):
-            result = await agent_client.CheckAgentHealthOnHost("host1").action()
-
-        self.assertFalse(result)
-        self.assertIn("does not support features: disks", result.message)
-
-    async def test_health_check_returns_false_on_http_error(self):
-        session = ClientSession(Response(status=500))
-
-        with mock.patch.object(agent_client.aiohttp, "ClientSession", lambda: session):
-            self.assertFalse(await agent_client.CheckAgentHealthOnHost("host1").action())
