+++ ydb/tools/mnc/cli/ut/ya.make
@@ -1,7 +1,13 @@
 PY3TEST()
 
 TEST_SRCS(
-    test_cli.py
+    test_agent_client.py
+    test_disks.py
+    test_main.py
+)
+
+PY_SRCS(
+    helpers.py
 )
 
 PEERDIR(
